backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
redis[hiredis]==5.0.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
Pillow==10.3.0
python-dotenv==1.0.1
